+++ app/index.html
@@ -5,7 +5,7 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/7.0.0/normalize.min.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
   <link rel="stylesheet" href="./css/noto/css/noto.css">
-  <link rel="stylesheet" href="./css/index.css">
+  <link rel="stylesheet" href="./index.css">
   <title>NOAH</title>
 </head>
 <body>
@@ -29,6 +29,6 @@
       <div id="test"></div>
     </div>
   </main>
-  <script src="./js/index.js"></script>
+  <script src="./index.js"></script>
 </body>
 </html>
